+++ CHANGELOG.md
@@ -2,6 +2,12 @@ Changelog for the gruf-prometheus gem.
 
 h3. Pending Release
 
+h3. 0.0.2
+
+- Have server handle sigints more gracefully
+- Cleaner starting of the gruf server and collectors
+- Improved logging and visibility around starting/stopping of collectors/server
+
 h3. 0.0.1
 
 - Initial public release